mise en place du glassmorphism
