+++ tailwind.config.js
@@ -4,6 +4,10 @@ module.exports = {
   theme: {
     extend: {
       colors: { pink: "#f38375" },
+      backgroundImage: {
+        index: "url('/src/Images/cool-background.png')",
+        leaderBoard: "url('/src/Images/leaderBoardBG.png')",
+      },
     },
   },
   plugins: [],
